+++ docs/source/create.rst
@@ -5,12 +5,14 @@ Create a start object
 Create an object that contains a DataFrame with at minimum one column that is the names of the files or folders in the supplied directory,
 
 .. code-block:: python
+    
+    import lrengine as lr
 
     path = "/path/to/directory/"
     lrobject = lr.start(path)
 
 
-You may define patterns to classify by. If a single pattern or list of patterns is given, the columns will be named according to the patterns and a bool will be supplied indicating the pattern was or was not found. This example adds the column "sample1" and puts True where found, False where not found,
+You may define patterns to classify by. If a single pattern or list of patterns is given, the columns will be named according to the patterns and a bool will be supplied indicating the pattern was or was not found. This example adds the column "sample1" and puts **True** where found, **False** where not found,
 
 .. code-block:: python
 
@@ -26,9 +28,9 @@ You may also use regular expressions to do more sophisticated pattern searches a
     lrobject = lr.start(path, patterns={"sample": "\d\d"})
 
 
-If this patterning is not found, bool (False) is entered. 
+If this patterning is not found, **False** is entered. 
 
-To mix these behaviors, add "bool" for the dict value and the column for the key is added with True or False added if the pattern is or is not found, respectively. This example makes a column for "sample" and classifies by any two digit number as above, but also adds a column called "blank_run" and classifies as True or False.
+To mix these behaviors, add "bool" for the dict value and the column for the key is added with **True** or **False** added if the pattern is or is not found, respectively. This example makes a column for "sample" and classifies by any two digit number as above, but also adds a column called "blank_run" and classifies as **True** or **False**.
 
 .. code-block:: python
