+++ docs/source/define_patterns.rst
@@ -2,9 +2,11 @@
 Defining patterns
 =================
 
-You may define patterns to classify by. If a single pattern or list of patterns is given, the columns will be named according to the patterns and a bool will be supplied indicating the pattern was or was not found. This example adds the column "sample1" and puts True where found, False where not found,
+You may define patterns to classify by. If a single pattern or list of patterns is given, the columns will be named according to the patterns and a bool will be supplied indicating the pattern was or was not found. This example adds the column "sample1" and puts **True** where found, **False** where not found,
 
 .. code-block:: python
+
+    import lrengine as lr
 
     path = "/path/to/directory/"
     lrobject = lr.start(path, patterns="sample1") # single pattern
@@ -18,11 +20,19 @@ You may also use regular expressions to do more sophisticated pattern searches a
     lrobject = lr.start(path, patterns={"sample": "\d\d"})
 
 
-If this patterning is not found, bool (False) is entered. 
+If this patterning is not found, **False** is entered. 
 
-To mix these behaviors, add "bool" for the dict value and the column for the key is added with True or False added if the pattern is or is not found, respectively. This example makes a column for "sample" and classifies by any two digit number as above, but also adds a column called "blank_run" and classifies as True or False.
+To mix these behaviors, add "bool" for the dict value and the column for the key is added with True or False added if the pattern is or is not found, respectively. This example makes a column for "sample" and classifies by any two digit number as above, but also adds a column called "blank_run" and classifies as **True** or **False**.
 
 .. code-block:: python
 
     path = "/path/to/directory/"
     lrobject = lr.start(path, patterns={"sample": "\d\d", "blank_run": bool})
+
+
+You can also modify an existing object to classify by different patterns using the **find_patterns()** method. For example, change the **patterns** then update the frame,
+
+.. code-block:: python
+
+    lrobject.patterns = {"sample": "\d", "sample": "\d\d"}
+    lrobject.find_patterns()
