+++ docs/source/intake.rst
@@ -5,6 +5,6 @@ intake
 ======
 
 .. autofunction:: lrengine.intake.date_injectors
-.. autofunction:: lrengine.intake.pattern_injectors
-.. autofunction:: lrengine.intake.names_filter
 .. autofunction:: lrengine.intake.dates_filter
+.. autofunction:: lrengine.intake.pattern_injectors
+.. autofunction:: lrengine.intake.patterns_filter
